Add more java code
Organize structure
