+++ c/exporecursion.c
@@ -4,25 +4,30 @@ float expo(int x,int y);
 main()
 {
 	int  x,y;
-	float res;
-	printf("enter x value:\n");
+	double res;
+	printf("enter x value: ");
 	scanf("%d",&x);
-	printf("enter y vlaue:\n");
+	printf("enter y vlaue: ");
 	scanf("%d",&y);
 	res = expo(x,y);
-	printf("%dth power of %d is %.2f ",y,x,res);
+	printf("%dth power of %d is %lf ",y,x,res);
 }
 float expo(int x, int y)
 {
-	float ans,neg;
-     
-     if (y<0){
-     	neg = pow(x,(-y));
-     	ans = 1/neg;
+	int i;
+	double result=1;
+     if(y<0){
+	   for(i=0;i>y;i--){
+		result*=x;
+		}
+		result=1/result;
     }
     else {
-    	ans = pow(x,y);
+    	for(i=0;i<y;i++)
+		{
+		result*=x;
+		}
 	}
-	return ans;
+	return result;
 }
