+++ java/Misc/Aligngui.java
