+++ java/Misc/Conversion.java
