+++ java/Misc/Dice.java
