+++ java/Misc/NewtonRaphsonMethod.java
