+++ java/Misc/ShapeIt.java
