+++ java/Misc/ZeroCrypto.java
